Title: 괄호 쌍 만들어주기, Time: 239ms, Memory: 14MB, Status: Passed - Codetree

+++ 250726/괄호 쌍 만들어주기/pair-parentheses.java
@@ -4,7 +4,7 @@ public class Main {
         Scanner sc = new Scanner(System.in);
         String A = sc.next();
         int len = A.length();
-        int[] lst = new int[len];
+        long[] lst = new long[len];
         for (int i = len - 2; i >= 0; i--){
             if (A.charAt(i) == ')' && A.charAt(i + 1) == ')')
                     lst[i] = lst[i + 1] + 1;
@@ -12,7 +12,7 @@ public class Main {
                 lst[i] = lst[i + 1];
         }
 
-        int res = 0;
+        long res = 0;
         for (int i = 1; i < len; i++){
             if (A.charAt(i) == '(' && A.charAt(i - 1) == '(')
                     res += lst[i];
